feat(ph-cmd): add new use command
