+++ clis/ph-cmd/src/commands/index.ts
@@ -2,19 +2,11 @@ import type { Command } from "commander";
 import { checkoutCommand } from "./checkout.js";
 import { helpCommand } from "./help.js";
 import { setupGlobalsCommand } from "./setup-globals.js";
-import { useCommand } from "./use.js";
 import { versionOption } from "./version.js";
-export {
-  checkoutCommand,
-  helpCommand,
-  setupGlobalsCommand,
-  useCommand,
-  versionOption,
-};
+export { checkoutCommand, helpCommand, setupGlobalsCommand, versionOption };
 export const commands = [
   setupGlobalsCommand,
   checkoutCommand,
-  useCommand,
   helpCommand,
   versionOption,
 ];
